Series formatting.
