+++ lib/key.mjs
@@ -2,6 +2,7 @@ import assert from 'assert'
 import binary_search from 'binary-search'
 import _ from 'underscore'
 
+import * as format from './format.mjs'
 import * as permutation from './permutation.mjs'
 
 /**
@@ -26,6 +27,10 @@ export class Key {
   getAt(idx) {
     // FIXME: Check range.
     return this.labels[idx]
+  }
+
+  getFormat(opts={}) {
+    return format.guess(this.labels)
   }
 }
 
@@ -70,5 +75,9 @@ export class IndexKey {
 
   getAt(idx) {
     return idx
+  }
+
+  getFormat(opts={}) {
+    return format.formatter(3)  // FIXME: Right number of digits.
   }
 }
