+++ lib/series.mjs
@@ -1,40 +1,40 @@
 import assert from 'assert'
 
 import { Key, IndexKey } from './key.mjs'
+import * as format from './format.mjs'
 import * as permutation from './permutation.mjs'
 
 export class Series {
-  constructor(labels, values) {
-    assert.equal(labels.length, values.length)
-    this.labels = labels
+  constructor(key, values) {
+    assert.equal(key.length, values.length)
+    this.key = key
     this.values = values
   }
 
   get(label) {
-    const idx = this.labels.get(label)
+    const idx = this.key.get(label)
     return this.values[idx]
   }
 
   get length() { return this.values.length }
-
-  // FIXME: Is this reasonable?
-  get key() { return this.labels instanceof Key ? this.labels : undefined }
-
   get summary() { return summarize(this.values) }
 
-  toString(labelFmt=FORMAT, fmt=FORMAT) {
+  toString() {
+    const keyFmt = this.key.getFormat()
+    const valFmt = format.guess(this.values)
     const lines = []
 
     // Header.
     lines.push(
-        format.palide('key', labelFmt.width) + ' '
-      + format.palide('value', fmt.width))
+        format.palide('key', keyFmt.width) + '   '
+      + format.palide('value', valFmt.width))
     // Underline.
-    lines.push('\u2501'.repeat(labelFmt.width) + '\u252d' + '\u2500'.repeat(fmt.width))
+    lines.push('\u2501'.repeat(keyFmt.width)
+      + '\u2501\u252d\u2500' + '\u2500'.repeat(valFmt.width))
 
     for (let i = 0; i < this.length; ++i)
       lines.push(
-        labelFmt(this.labels.getAt(i)) + '\u2502' + fmt(this.values[i]))
+        keyFmt(this.key.getAt(i)) + ' \u2502 ' + valFmt(this.values[i]))
 
     lines.push('')
     return lines.join('\n')
